move properties to own file, change transfer properties to use PAPI types
